Move map component into sub folder
